Created the basic playfield
A plane strip for the player to roll on with two walls on either side, a sphere for the player and a cube prefab for the obstacle. Also added three spawnpoints (prefab) for the obstacles and an end zone. The end zone is a cube collider which triggers the destruction of an obstacle when they come into contact.